Use uint16_t for vehicle model IDs instead of uint32_t
Model IDs are unsigned, so sign extending them into the resolver calls
was wrong even if harmless today. Switched every naked hook to a zero
extending load instead, same instruction length so nothing else changes.

+++ Client/mods/deathmatch/logic/CClientVehicle.cpp
@@ -1362,7 +1362,7 @@ void CClientVehicle::SetLandingGearDown(bool bLandingGearDown)
 // since SetModelBlocking calls in here before that pointer has been updated.
 static VehicleType GetVehicleTypeForModel(unsigned short usModel)
 {
-    std::uint32_t ulModel = usModel;
+    std::uint16_t ulModel = usModel;
     if (ulModel < 400 || ulModel > 611)
     {
         if (CModelInfo* pModelInfo = g_pGame->GetModelInfo(ulModel))
